separated run-specific directories into run directory; added generate directory
